moved tests into separate directory + added some more
