+++ src/lib.rs
@@ -15,7 +15,7 @@ use crate::error::IpApiError;
 
 #[cfg(feature = "blocking")]
 pub mod blocking;
-mod client;
+pub mod client;
 pub mod constant;
 pub mod error;
 pub mod model;
@@ -90,166 +90,5 @@ impl AsyncIpApi for IpApiClient {
 
     fn get_http_client(&self) -> &Client {
         &self.client
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use crate::blocking::client::BlockingIpApiClient;
-    use crate::client::{AsyncIpApi, BlockingIpApi};
-    use crate::error::IpApiError;
-    use crate::model::ip_response::{IpDefaultResponse, IpFullResponse};
-    use crate::util::urls::{build_http_url_from_struct, build_https_url_from_struct, build_url_without_fields};
-    use crate::IpApiClient;
-
-    const TEST_IP: &str = "127.0.0.1";
-    const EXTERN_TEST_IP: &str = "8.8.8.8";
-
-    #[test]
-    fn test_default_http_url() {
-        let url = build_url_without_fields(false, &TEST_IP.to_string());
-        assert_eq!(url, format!("http://ip-api.com/json/{}", TEST_IP))
-    }
-
-    #[test]
-    fn test_default_https_url() {
-        let url = build_url_without_fields(true, &TEST_IP.to_string());
-        assert_eq!(url, format!("https://ip-api.com/json/{}", TEST_IP))
-    }
-
-    #[test]
-    fn test_custom_http_url() {
-        let url = build_http_url_from_struct::<IpFullResponse>(&TEST_IP.to_string());
-        assert_eq!(url, format!("http://ip-api.com/json/{}?fields=query,status,continent,continentCode,country,countryCode,region,regionName,city,district,zip,lat,lon,timezone,offset,currency,isp,org,as,asname,reverse,mobile,proxy,hosting,message", TEST_IP));
-    }
-
-    #[test]
-    fn test_custom_https_url() {
-        let url = build_https_url_from_struct::<IpFullResponse>(&TEST_IP.to_string());
-        assert_eq!(url, format!("https://ip-api.com/json/{}?fields=query,status,continent,continentCode,country,countryCode,region,regionName,city,district,zip,lat,lon,timezone,offset,currency,isp,org,as,asname,reverse,mobile,proxy,hosting,message", TEST_IP));
-    }
-
-    #[tokio::test]
-    async fn test_client() {
-        let client = IpApiClient::new();
-        assert_eq!(client.api_key, None);
-        let result = client.query_api_default(EXTERN_TEST_IP).await;
-        let expected = IpDefaultResponse {
-            query: EXTERN_TEST_IP.to_string(),
-            status: "success".to_string(),
-            country: "United States".to_string(),
-            country_code: "US".to_string(),
-            region: "VA".to_string(),
-            region_name: "Virginia".to_string(),
-            city: "Ashburn".to_string(),
-            zip: "20149".to_string(),
-            lat: 39.03,
-            lon: -77.5,
-            timezone: "America/New_York".to_string(),
-            isp: "Google LLC".to_string(),
-            org: "Google Public DNS".to_string(),
-            asn: "AS15169 Google LLC".to_string(),
-        };
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_client_custom() {
-        let client = IpApiClient::new();
-        assert_eq!(client.api_key, None);
-        let result = client.query_api::<IpDefaultResponse>(EXTERN_TEST_IP).await;
-        let expected = IpDefaultResponse {
-            query: EXTERN_TEST_IP.to_string(),
-            status: "success".to_string(),
-            country: "United States".to_string(),
-            country_code: "US".to_string(),
-            region: "VA".to_string(),
-            region_name: "Virginia".to_string(),
-            city: "Ashburn".to_string(),
-            zip: "20149".to_string(),
-            lat: 39.03,
-            lon: -77.5,
-            timezone: "America/New_York".to_string(),
-            isp: "Google LLC".to_string(),
-            org: "Google Public DNS".to_string(),
-            asn: "AS15169 Google LLC".to_string(),
-        };
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), expected);
-    }
-
-    #[tokio::test]
-    async fn test_error_reserved_range() {
-        let client = IpApiClient::new();
-        let result = client.query_api::<IpDefaultResponse>("127.0.0.1").await;
-        match result.err().unwrap() {
-            IpApiError::ReservedRange(error_response) => {
-                assert_eq!(error_response.message, "reserved range");
-            }
-            _ => panic!("Wrong error type returned."),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_error_invalid_query() {
-        let client = IpApiClient::new();
-        let result = client.query_api::<IpDefaultResponse>("Invalid Query").await;
-        match result.err().unwrap() {
-            IpApiError::InvalidQuery(error_response) => {
-                assert_eq!(error_response.message, "invalid query");
-            }
-            _ => panic!("Wrong error type returned."),
-        }
-    }
-
-    #[test]
-    fn test_blocking_client() {
-        let client = BlockingIpApiClient::new();
-        assert_eq!(client.api_key, None);
-        let result = client.query_api_default(EXTERN_TEST_IP);
-        let expected = IpDefaultResponse {
-            query: EXTERN_TEST_IP.to_string(),
-            status: "success".to_string(),
-            country: "United States".to_string(),
-            country_code: "US".to_string(),
-            region: "VA".to_string(),
-            region_name: "Virginia".to_string(),
-            city: "Ashburn".to_string(),
-            zip: "20149".to_string(),
-            lat: 39.03,
-            lon: -77.5,
-            timezone: "America/New_York".to_string(),
-            isp: "Google LLC".to_string(),
-            org: "Google Public DNS".to_string(),
-            asn: "AS15169 Google LLC".to_string(),
-        };
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), expected);
-    }
-
-    #[test]
-    fn test_blocking_client_custom() {
-        let client = BlockingIpApiClient::new();
-        assert_eq!(client.api_key, None);
-        let result = client.query_api::<IpDefaultResponse>(EXTERN_TEST_IP);
-        let expected = IpDefaultResponse {
-            query: EXTERN_TEST_IP.to_string(),
-            status: "success".to_string(),
-            country: "United States".to_string(),
-            country_code: "US".to_string(),
-            region: "VA".to_string(),
-            region_name: "Virginia".to_string(),
-            city: "Ashburn".to_string(),
-            zip: "20149".to_string(),
-            lat: 39.03,
-            lon: -77.5,
-            timezone: "America/New_York".to_string(),
-            isp: "Google LLC".to_string(),
-            org: "Google Public DNS".to_string(),
-            asn: "AS15169 Google LLC".to_string(),
-        };
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), expected);
     }
 }
