+++ tests/constants.rs
@@ -1,0 +1,2 @@
+pub const TEST_IP: &str = "127.0.0.1";
+pub const EXTERN_TEST_IP: &str = "8.8.8.8";
