+++ tests/end2end/client/mod.rs
@@ -1,0 +1,3 @@
+pub mod async_client_test;
+pub mod blocking_client_test;
+pub mod utils;
